Final project completed

+++ Week5/final_project/Logic.py
@@ -1,5 +1,5 @@
 import Service
-
+import Objects
 
 class GameEngine:
     objects = []
@@ -11,6 +11,10 @@ class GameEngine:
     score = 0.
     game_process = True
     show_help = False
+    sprite_size = None
+    is_dead = False
+
+    show_start = True
 
     def subscribe(self, obj):
         self.subscribers.add(obj)
@@ -30,29 +34,41 @@ class GameEngine:
     def interact(self):
         for obj in self.objects:
             if list(obj.position) == self.hero.position:
-                self.delete_object(obj)
-                obj.interact(self, self.hero)
+                
+                if type(obj) != Objects.Enemy:
+                    self.delete_object(obj)     
+                    obj.interact(self)
+                else:
+                    obj.interact(self)
+                    if obj.hp == 0:
+                        self.delete_object(obj)
+                    if self.hero.hp == 0:
+                        self.notify("You are dead!!")
+                        self.is_dead = True
+                        self = Service.reload_game(self)
+
+                #
 
     # MOVEMENT
     def move_up(self):
         self.score -= 0.02
         if self.map[self.hero.position[1] - 1][self.hero.position[0]] == Service.wall:
             return
-        self.hero.position[1] -= 1
+        self.hero.position[1] -= 1 
         self.interact()
 
     def move_down(self):
         self.score -= 0.02
         if self.map[self.hero.position[1] + 1][self.hero.position[0]] == Service.wall:
             return
-        self.hero.position[1] += 1
+        self.hero.position[1] += 1  
         self.interact()
 
     def move_left(self):
         self.score -= 0.02
         if self.map[self.hero.position[1]][self.hero.position[0] - 1] == Service.wall:
             return
-        self.hero.position[0] -= 1
+        self.hero.position[0] -= 1 
         self.interact()
 
     def move_right(self):
